feat: make carrousel on status money component
