+++ package.json
@@ -28,6 +28,7 @@
     "date-fns": "2.28.0",
     "dayjs": "^1.11.6",
     "formik": "^2.2.9",
+    "keen-slider": "^6.8.6",
     "moment": "^2.29.4",
     "phosphor-react": "^1.4.1",
     "react": "^18.0.0",
